feat: move imgproc to internal library
