+++ average_test.go
@@ -7,7 +7,6 @@ import (
 
 	. "github.com/ajdnik/imghash"
 	"github.com/ajdnik/imghash/hashtype"
-	"github.com/ajdnik/imghash/imgproc"
 	"github.com/ajdnik/imghash/similarity"
 )
 
@@ -30,7 +29,7 @@ func TestAverage_Calculate(t *testing.T) {
 	for _, tt := range averageCalculateTests {
 		t.Run(tt.filename, func(t *testing.T) {
 			hash := NewAverage(WithSize(tt.width, tt.height), WithInterpolation(tt.resizeType))
-			img, _ := imgproc.Read(tt.filename)
+			img, _ := OpenImage(tt.filename)
 			result, err := hash.Calculate(img)
 			if err != nil {
 				t.Fatalf("unexpected error: %v", err)
@@ -45,7 +44,7 @@ func TestAverage_Calculate(t *testing.T) {
 
 func ExampleAverage_Calculate() {
 	// Read image from file
-	img, _ := imgproc.Read("assets/cat.jpg")
+	img, _ := OpenImage("assets/cat.jpg")
 	// Create new Average Hash using default parameters
 	avg := NewAverage()
 	// Calculate hash
@@ -74,8 +73,8 @@ func TestAverage_Distance(t *testing.T) {
 	for _, tt := range averageDistanceTests {
 		t.Run(fmt.Sprintf("%v %v", tt.firstImage, tt.secondImage), func(t *testing.T) {
 			hash := NewAverage(WithSize(tt.width, tt.height), WithInterpolation(tt.resizeType))
-			img1, _ := imgproc.Read(tt.firstImage)
-			img2, _ := imgproc.Read(tt.secondImage)
+			img1, _ := OpenImage(tt.firstImage)
+			img2, _ := OpenImage(tt.secondImage)
 			h1, _ := hash.Calculate(img1)
 			h2, _ := hash.Calculate(img2)
 			dist, _ := similarity.Hamming(h1, h2)
