+++ blockmean_test.go
@@ -7,7 +7,6 @@ import (
 
 	. "github.com/ajdnik/imghash"
 	"github.com/ajdnik/imghash/hashtype"
-	"github.com/ajdnik/imghash/imgproc"
 	"github.com/ajdnik/imghash/similarity"
 )
 
@@ -39,7 +38,7 @@ func TestBlockMean_Calculate(t *testing.T) {
 	for _, tt := range blockMeanCalculateTests {
 		t.Run(tt.filename, func(t *testing.T) {
 			hash := NewBlockMean(WithSize(tt.width, tt.height), WithInterpolation(tt.resizeType), WithBlockSize(tt.blockWidth, tt.blockHeight), WithBlockMeanMethod(tt.method))
-			img, _ := imgproc.Read(tt.filename)
+			img, _ := OpenImage(tt.filename)
 			result, err := hash.Calculate(img)
 			if err != nil {
 				t.Fatalf("unexpected error: %v", err)
@@ -54,7 +53,7 @@ func TestBlockMean_Calculate(t *testing.T) {
 
 func ExampleBlockMean_Calculate() {
 	// Read image from file
-	img, _ := imgproc.Read("assets/cat.jpg")
+	img, _ := OpenImage("assets/cat.jpg")
 	// Create new Block Mean Hash using default parameters
 	block := NewBlockMean()
 	// Calculate hash
@@ -91,8 +90,8 @@ func TestBlockMean_Distance(t *testing.T) {
 	for _, tt := range blockMeanDistanceTests {
 		t.Run(fmt.Sprintf("%v %v", tt.firstImage, tt.secondImage), func(t *testing.T) {
 			hash := NewBlockMean(WithSize(tt.width, tt.height), WithInterpolation(tt.resizeType), WithBlockSize(tt.blockWidth, tt.blockHeight), WithBlockMeanMethod(tt.method))
-			img1, _ := imgproc.Read(tt.firstImage)
-			img2, _ := imgproc.Read(tt.secondImage)
+			img1, _ := OpenImage(tt.firstImage)
+			img2, _ := OpenImage(tt.secondImage)
 			h1, _ := hash.Calculate(img1)
 			h2, _ := hash.Calculate(img2)
 			dist, _ := similarity.Hamming(h1, h2)
