+++ colormoment.go
@@ -4,7 +4,7 @@ import (
 	"image"
 
 	"github.com/ajdnik/imghash/hashtype"
-	"github.com/ajdnik/imghash/imgproc"
+	"github.com/ajdnik/imghash/internal/imgproc"
 )
 
 // ColorMoment is a perceptual hash that uses the method described in
