+++ difference_test.go
@@ -9,7 +9,6 @@ import (
 
 	. "github.com/ajdnik/imghash"
 	"github.com/ajdnik/imghash/hashtype"
-	"github.com/ajdnik/imghash/imgproc"
 	"github.com/ajdnik/imghash/similarity"
 )
 
@@ -32,7 +31,7 @@ func TestDifference_Calculate(t *testing.T) {
 	for _, tt := range differenceCalculateTests {
 		t.Run(tt.filename, func(t *testing.T) {
 			hash := NewDifference(WithSize(tt.width, tt.height), WithInterpolation(tt.resizeType))
-			img, _ := imgproc.Read(tt.filename)
+			img, _ := OpenImage(tt.filename)
 			result, err := hash.Calculate(img)
 			if err != nil {
 				t.Fatalf("unexpected error: %v", err)
@@ -47,7 +46,7 @@ func TestDifference_Calculate(t *testing.T) {
 
 func ExampleDifference_Calculate() {
 	// Read image from file
-	img, _ := imgproc.Read("assets/cat.jpg")
+	img, _ := OpenImage("assets/cat.jpg")
 	// Create new Difference Hash using default parameters
 	diff := NewDifference()
 	// Calculate hash
@@ -76,8 +75,8 @@ func TestDifference_Distance(t *testing.T) {
 	for _, tt := range differenceDistanceTests {
 		t.Run(fmt.Sprintf("%v %v", tt.firstImage, tt.secondImage), func(t *testing.T) {
 			hash := NewDifference(WithSize(tt.width, tt.height), WithInterpolation(tt.resizeType))
-			img1, _ := imgproc.Read(tt.firstImage)
-			img2, _ := imgproc.Read(tt.secondImage)
+			img1, _ := OpenImage(tt.firstImage)
+			img2, _ := OpenImage(tt.secondImage)
 			h1, _ := hash.Calculate(img1)
 			h2, _ := hash.Calculate(img2)
 			dist, _ := similarity.Hamming(h1, h2)
