+++ internal/imgproc/common.go
