+++ internal/imgproc/equalize.go
