+++ internal/imgproc/filter.go
