+++ internal/imgproc/moments.go
