+++ internal/imgproc/read.go
